~ More eager cancellation check in StateFlow, better test
